+++ src/gpu/graphite/vk/VulkanCaps.cpp
@@ -1568,8 +1568,8 @@ void VulkanCaps::buildKeyForTexture(SkISize dimensions,
     SkASSERT(vkSpec.fImageUsageFlags            < (1u << 12)); // imageUsageFlags are bits 20-32
 
     // We need two uint32_ts for dimensions, 1 for format, and 2 for the rest of the information.
-    static constexpr int kNum32DataCntNoYcbcr =  2 + 1 + 2;
-    int num32DataCnt = kNum32DataCntNoYcbcr;
+    static constexpr uint16_t kNum32DataCntNoYcbcr =  2 + 1 + 2;
+    uint16_t num32DataCnt = kNum32DataCntNoYcbcr;
 
     // If a texture w/ an external format is being used, that information must also be appended.
     const VulkanYcbcrConversionInfo& ycbcrInfo = TextureInfos::GetVulkanYcbcrConversionInfo(info);
